Lag komponent for forside (#76)
### Hvorfor er dette nødvendig? ✨
For at alle stønadsinngangene skal kunne føre til en forside som er
relativt lik bortsett fra teksten skal det lages en komponent for
forsiden som kan gjenbrukes.

[Lenke til trello
kort](https://trello.com/c/XLkwb8zb/122-egen-komponent-for-forside)

### Hvordan er det løst? 🧠
Mappen `sider` er lagt til i `app` og det er laget en ny komponent
`Forside.tsx` i denne mappen. Innholdet fra `_index.tsx` er flyttet inn
i denne nye komponenten. I tillegg er rekkefølgen på const-ene og
funksjonene litt endret, og const for `hentBrukerHilsen` fjernet og satt
direkte inn i overskrift propen til `VeilderPanel`
